complete the calendar design
