+++ src/component/ProfileTrainer/Calendar/styles.css
@@ -1,11 +1,13 @@
 .example-container {
+  width: 100%;
   /* padding-top: 4rem; */
 }
 
 .calendar {
   margin: auto;
-  max-width: 60rem;
-  border: 1px solid #e3e3e3;
+  width: 100%;
+  max-width: 994px;
+  box-shadow: 0 0px 4px rgba(0, 0, 0, 0.24);
   border-radius: 40px;
 }
 
@@ -13,13 +15,13 @@
   display: flex;
   gap: 1rem;
   border-bottom: 1px solid #e3e3e3;
-  padding: 0.75rem 1.25rem;
+  padding: 0.9rem 3rem;
 }
 
 .weekdays {
   display: grid;
   grid-template-columns: repeat(7, 1fr);
-  padding: 0.5rem 1.25rem;
+  padding: 0.9rem 1.25rem;
   text-align: center;
   font-weight: 600;
   color: #3e4678;
@@ -28,24 +30,29 @@
 
 .weekday {
   color: #3e4678;
-  font-size: 0.875rem;
+  font-size: 1.25rem;
+  font-weight: 800;
 }
 
 .days {
   display: grid;
   grid-template-columns: repeat(7, 1fr);
-  padding: 0.5rem 1.25rem;
-  row-gap: 5px;
+  padding: 0.9rem 1.25rem;
+  row-gap: 14px;
 }
 
 .day {
-  padding: 0.375rem;
+  /* padding: 0.375rem; */
+  max-width: 127px;
+  height: 94px;
 }
 
 .day-title {
   width: 100%;
   font-size: 0.875rem;
-  font-weight: 600;
+  text-transform: uppercase;
+  padding: 0 0.5rem;
+  font-weight: normal;
   white-space: nowrap;
   overflow: hidden;
   text-overflow: ellipsis;
@@ -76,8 +83,8 @@
 
 .day-button {
   margin: auto;
-  height: 2.5rem;
   width: 2.5rem;
+  height: 2.5rem;
   font-size: 0.875rem;
   border-radius: 50%;
   display: flex;
@@ -88,8 +95,11 @@
 }
 
 .dot-container {
-  width: 0.3rem;
-  height: 0.3rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 0.6rem;
+  height: 0.6rem;
   margin: 0 auto;
   margin-top: 0.25rem;
 }
@@ -192,28 +202,6 @@
 
 .active {
   background-color: #f3f4f6;
-}
-
-.navigation {
-  display: flex;
-  align-items: center;
-  gap: 0.3rem;
-}
-
-.nav-button {
-  background: none;
-  border: none;
-  cursor: pointer;
-}
-
-.nav-date {
-  font-weight: bold;
-  font-size: 0.875rem;
-}
-
-.nav-icon {
-  width: 0.9rem;
-  height: 0.9rem;
 }
 
 .sr-only {
