+++ src/component/ProfileTrainer/styles.module.css
@@ -71,7 +71,8 @@
 
 .main {
   display: flex;
-  gap: 5.5rem;
+  flex-wrap: wrap;
+  gap: 1.25rem;
   margin: 5rem 0px;
 }
 
@@ -84,6 +85,7 @@
 
 .main .sideBar {
   flex: 1;
+  max-width: 452px;
 }
 
 .sideBar {
@@ -91,10 +93,10 @@
   height: 100%;
   border-radius: 2.5rem;
   box-shadow: 0px 0px 4px var(--shadow-color);
-  padding: 54px;
+  padding: 35px 62px;
   display: flex;
   flex-direction: column;
-  gap: 50px;
+  gap: 24px;
 }
 
 .sideBarContainer {
